scripts: lunarvim: make install idempotent

+++ labvm/scripts/files/home/lunarvim/install.sh
@@ -1,11 +1,15 @@
 #!/bin/bash
+# Local LunarVim Installation script
 
 python3 -m pip install --user pynvim
 
-# Local LunarVim Installation script
-rm -rf "$HOME/.local/share/lunarvim" "$HOME/.config/lvim" "$HOME/.cache/lvim"
+if [[ "$DEBUG" -gt 1 ]]; then
+	rm -rf "$HOME/.local/share/lunarvim" "$HOME/.config/lvim" "$HOME/.cache/lvim"
+fi
 
-bash <(curl -s \
-	https://raw.githubusercontent.com/lunarvim/lunarvim/master/utils/installer/install.sh) \
-	--yes --no-install-dependencies
+if [[ ! -d "$HOME/.local/share/lunarvim" ]]; then
+	bash <(curl -s \
+		https://raw.githubusercontent.com/lunarvim/lunarvim/master/utils/installer/install.sh) \
+		--yes --no-install-dependencies
+fi
